+++ languagetool-standalone/src/main/java/org/languagetool/gui/LanguageToolSupport.java
@@ -851,7 +851,17 @@ class LanguageToolSupport {
           Color ulColor = config.getUnderlineColor(span.rule.getCategory().getName());
           Color colorForIssueType = getConfig().getErrorColors().get(issueType);
           Color bgColor = colorForIssueType != null ? colorForIssueType : null;
-          Color underlineColor = ITSIssueType.Misspelling == span.rule.getLocQualityIssueType() ? Color.red : ulColor;
+          Color underlineColor = null;
+          if (ITSIssueType.Typographical == span.rule.getLocQualityIssueType()) {
+           underlineColor = Color.cyan;
+           }
+          else if (ITSIssueType.Duplication == span.rule.getLocQualityIssueType()) {
+           underlineColor = Color.gray;
+           }
+          else if (ITSIssueType.Misspelling == span.rule.getLocQualityIssueType()) {
+           underlineColor = Color.red;
+           }
+          else underlineColor = ulColor;
           HighlightPainter painter = new HighlightPainter(bgColor, underlineColor);
           h.addHighlight(span.start, span.end, painter);
         }